refactor(collection-ui): move main menu group
