+++ apps/frontend/src/app/app.component.spec.ts
@@ -3,7 +3,7 @@ import { TestBed } from '@angular/core/testing';
 import { provideRouter } from '@angular/router';
 import { AppComponent } from './app.component';
 import { AppMainMenuComponent } from './components/app-main-menu/app-main-menu.component';
-import { UiDesktopModule } from './modules/ui-desktop/ui-desktop.module';
+import { CkadUiDesktopModule } from './modules/ckapps-design-system/ckad-ui-desktop/ckad-ui-desktop.module';
 
 @Component({
   selector: 'ckad-desktop-window-title-bar',
@@ -27,7 +27,7 @@ describe('AppComponent', () => {
 
     TestBed.overrideComponent(AppComponent, {
       remove: {
-        imports: [AppMainMenuComponent, UiDesktopModule],
+        imports: [AppMainMenuComponent, CkadUiDesktopModule],
       },
       add: {
         imports: [MockAppMainMenuComponent, MockWindowTitleBarComponent],
