+++ apps/frontend/src/app/components/app-main-menu/app-main-menu.component.html
@@ -43,5 +43,5 @@
     </mat-nav-list>
   </ckad-main-menu-group>
 
-  <overckd-collections-main-menu-group [collections]="collections()" />
+  <overckd-collection-main-menu-group [collections]="collections()" />
 </ckad-main-menu>
