+++ apps/frontend/src/app/modules/domain/domain.module.ts
@@ -1,11 +1,8 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
-// Domain modules
-import { RecipeModule } from './recipe/recipe.module';
-import { RecipeCollectionModule } from './recipe-collection/recipe-collection.module';
+import { NgModule } from '@angular/core';
 import { IngredientModule } from './ingredient/ingredient.module';
 import { PortionModule } from './portion/portion.module';
+import { RecipeModule } from './recipe/recipe.module';
 import { TagModule } from './tag/tag.module';
 
 @NgModule({
@@ -13,16 +10,10 @@ import { TagModule } from './tag/tag.module';
   imports: [
     CommonModule,
     RecipeModule,
-    RecipeCollectionModule,
     IngredientModule,
     PortionModule,
     TagModule,
   ],
-  exports: [
-    RecipeModule,
-    RecipeCollectionModule,
-    IngredientModule,
-    PortionModule,
-  ],
+  exports: [RecipeModule, IngredientModule, PortionModule],
 })
 export class DomainModule {}
