--- apps/frontend/src/app/modules/domain/recipe-collection/recipe-collection.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CollectionsMainMenuGroupComponent } from './components/collections-main-menu-group/collections-main-menu-group.component';
-
-@NgModule({
-  imports: [CollectionsMainMenuGroupComponent],
-  exports: [CollectionsMainMenuGroupComponent],
-})
-export class RecipeCollectionModule {}
