+++ libs/shared/collection/ui/src/index.ts
@@ -1,1 +1,2 @@
-export * from './lib/recipe-collection.component';
+export * from './lib/collection-main-menu-group.component';
+export * from './lib/collection.component';
