+++ libs/shared/collection/ui/src/lib/collection-main-menu-group.component.html
